"Reselect" implementation to memoize derived data.
