+++ package.json
@@ -177,9 +177,10 @@
         "node-sass": "^4.9.2",
         "openseadragon": "^2.3.1",
         "qrcode-generator": "^1.3.1",
-        "react": "16.4.1",
+        "re-reselect": "^2.1.0",
+        "react": "16.4.2",
         "react-a11y": "1.0.0",
-        "react-dom": "16.4.1",
+        "react-dom": "16.4.2",
         "react-hot-loader": "4.3.3",
         "react-motion": "^0.5.2",
         "react-password-strength": "^2.3.1",
@@ -194,6 +195,7 @@
         "redux-saga": "0.16.0",
         "redux-saga-test-plan": "^3.3.1",
         "regenerator-runtime": "^0.11.0",
+        "reselect": "^3.0.1",
         "sass-loader": "^6.0.6",
         "sinon": "^1.17.7",
         "style-loader": "^0.13.1",
